# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RAG search API for book publishing archive
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type BookStatus = typeof BookStatus[keyof typeof BookStatus];


export const BookStatus = {
  pending: 'pending',
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface Book {
  id: number;
  title: string;
  author: string;
  /** @nullable */
  publishedYear: number | null;
  /** @nullable */
  description: string | null;
  /** Server-side path to the source file (PDF or plain text) */
  filePath: string;
  status: BookStatus;
  /** @nullable */
  errorMessage?: string | null;
  totalChunks: number;
  createdAt: string;
  updatedAt: string;
}

export interface BookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  author: string;
  /** @nullable */
  publishedYear?: number | null;
  /** @nullable */
  description?: string | null;
  /**
     * Server-side absolute path to the PDF or text file
     * @minLength 1
     */
  filePath: string;
}

export interface BookUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  author?: string;
  /** @nullable */
  publishedYear?: number | null;
  /** @nullable */
  description?: string | null;
}

export interface BookListPage {
  books: Book[];
  total: number;
}

export interface Chunk {
  id: number;
  bookId: number;
  chunkIndex: number;
  content: string;
  /** @nullable */
  chapterTitle: string | null;
  /** @nullable */
  pageStart: number | null;
  /** @nullable */
  pageEnd: number | null;
  createdAt: string;
}

export interface ChunkListPage {
  chunks: Chunk[];
  total: number;
}

export interface SearchInput {
  /**
     * Plain-English question to ask against the book archive
     * @minLength 1
     */
  query: string;
}

export interface Citation {
  bookId: number;
  bookTitle: string;
  author: string;
  /** @nullable */
  chapterTitle?: string | null;
  /** @nullable */
  pageStart?: number | null;
  /** @nullable */
  pageEnd?: number | null;
  excerpt: string;
  relevanceScore: number;
}

export interface SearchResult {
  id: number;
  queryText: string;
  answer: string;
  citations: Citation[];
  /** Clerk user ID */
  userId: string;
  createdAt: string;
}

export interface SearchHistoryPage {
  results: SearchResult[];
  total: number;
}

export interface CatalogStats {
  totalBooks: number;
  booksReady: number;
  booksProcessing: number;
  booksError: number;
  totalChunks: number;
  totalSearches: number;
}

export type ListSearchHistoryParams = {
limit?: number;
offset?: number;
};

export type ListRecentSearchesParams = {
limit?: number;
};

export type ListBooksParams = {
status?: ListBooksStatus;
limit?: number;
offset?: number;
};

export type ListBooksStatus = typeof ListBooksStatus[keyof typeof ListBooksStatus];


export const ListBooksStatus = {
  pending: 'pending',
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;