# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RAG search API for book publishing archive
 * OpenAPI spec version: 0.1.0
 */

export interface BookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  author: string;
  /** @nullable */
  publishedYear?: number | null;
  /** @nullable */
  description?: string | null;
  /**
     * Server-side absolute path to the PDF or text file
     * @minLength 1
     */
  filePath: string;
}