# lib/api-zod/src/generated/types/catalogStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RAG search API for book publishing archive
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogStats {
  totalBooks: number;
  booksReady: number;
  booksProcessing: number;
  booksError: number;
  totalChunks: number;
  totalSearches: number;
}