# lib/api-zod/src/generated/types/citation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RAG search API for book publishing archive
 * OpenAPI spec version: 0.1.0
 */

export interface Citation {
  bookId: number;
  bookTitle: string;
  author: string;
  /** @nullable */
  chapterTitle?: string | null;
  /** @nullable */
  pageStart?: number | null;
  /** @nullable */
  pageEnd?: number | null;
  excerpt: string;
  relevanceScore: number;
}