# lib/api-zod/src/generated/types/searchInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RAG search API for book publishing archive
 * OpenAPI spec version: 0.1.0
 */

export interface SearchInput {
  /**
     * Plain-English question to ask against the book archive
     * @minLength 1
     */
  query: string;
}