# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RAG search API for book publishing archive
 * OpenAPI spec version: 0.1.0
 */
import type { Citation } from './citation';

export interface SearchResult {
  id: number;
  queryText: string;
  answer: string;
  citations: Citation[];
  /** Clerk user ID */
  userId: string;
  createdAt: Date;
}